Controller Test including Usercontroller test and Taxicontroller test

+++ src/main/java/com/example/taxibillingsystem/contract/request/AmountBalanceRequest.java
@@ -10,6 +10,5 @@ import lombok.Getter;
 @Builder
 @ValidUserBalance
 public class AmountBalanceRequest {
-
-    private int accountBalance;
+    private long accountBalance;
 }
